Make deploy compile SCSS

+++ deploy.py
@@ -8,29 +8,30 @@ if len(sys.argv) < 2:
     sys.exit()
 sitename = sys.argv[1]
 
-# Remove contents of remote source directory
-subprocess.call(
- "ssh {} 'rm -r ~/{}/source/* >& /dev/null'".format(sitename, sitename), shell=True
-)
-
-# What files are being tracked by git?
-tracked_files = subprocess.check_output("git ls-files", shell=True).decode()
-tracked_files = list(filter(bool, tracked_files.split("\n")))
-
-# Push source code to remote
-commands = []
-for file_ in tracked_files:
-    directory = "/".join(file_.split("/")[:-1])
-    commands.append("mkdir -p ~/{}/source/{}".format(sitename, directory))
-commands = set(commands)
-subprocess.call(
- "ssh {} '{}'".format(sitename, ";".join(commands)), shell=True
-)
-files = []
-for file_ in tracked_files:
+if len(sys.argv) < 3 or sys.argv[2] != "nofiles":
+    # Remove contents of remote source directory
     subprocess.call(
-     "scp -r ./{} {}:~/{}/source/{}".format(file_, sitename, sitename, file_), shell=True
+     "ssh {} 'rm -r ~/{}/source/* >& /dev/null'".format(sitename, sitename), shell=True
     )
+
+    # What files are being tracked by git?
+    tracked_files = subprocess.check_output("git ls-files", shell=True).decode()
+    tracked_files = list(filter(bool, tracked_files.split("\n")))
+
+    # Push source code to remote
+    commands = []
+    for file_ in tracked_files:
+        directory = "/".join(file_.split("/")[:-1])
+        commands.append("mkdir -p ~/{}/source/{}".format(sitename, directory))
+    commands = set(commands)
+    subprocess.call(
+     "ssh {} '{}'".format(sitename, ";".join(commands)), shell=True
+    )
+    files = []
+    for file_ in tracked_files:
+        subprocess.call(
+         "scp -r ./{} {}:~/{}/source/{}".format(file_, sitename, sitename, file_), shell=True
+        )
 
 # Turn off Debug
 subprocess.call(
@@ -74,7 +75,7 @@ else:
 
 # Install pip packages
 subprocess.call(
- "ssh {} '~/{}/env/bin/pip install -r ~/{}/source/requirements.txt -r ~/{}/source/requirements-live.txt'".format(sitename, sitename, sitename),
+ "ssh {} '~/{}/env/bin/pip install -r ~/{}/source/requirements.txt -r ~/{}/source/requirements-live.txt'".format(sitename, sitename, sitename, sitename),
  shell=True
 )
 
@@ -85,6 +86,10 @@ subprocess.call(
 )
 
 # Deploy static files
+subprocess.call(
+ "ssh {} 'cd ~/{}/source && ../env/bin/python manage.py compilescss'".format(sitename, sitename),
+ shell=True
+)
 subprocess.call(
  "ssh {} 'cd ~/{}/source && ../env/bin/python manage.py collectstatic --noinput'".format(sitename, sitename),
  shell=True
